Move serde around.
